Major refactoring and cleanup
The main QML data model is now only an integer. And the data
is now stored in a central javascript datastructure, which
gets incrementally populated.
Before, the QML data model was populated with 50 exercises
when starting a lesson. That was to slow on a phone for too
few excercises.

Also the new strucure will allow a mixed lesson with mixed
exercises from all other available lessons.