Launcher icon in the new "Anna"-look
With rounded background shape.

Of course I had to create the .svg from scratch, myself[*].
There is tons of information about the new launcher icons
on developer.nokia.com. Even template icons for download,
but these are not available in the obvious .svg format
but instead as Adobe Indesign (expensive software) files.

That seems bizarre, but after all this is consistent
Symbian developer experience ;)

[*] Of course, I did not create it from scratch ;)
I rather stole it from the awesome TwimGo and tweaked it.